refactor: rename link parameter with assign parameter and relative components
